Fixing compiler warnings
All but one warning was for format specifiers (-WFormat) in sprintf().
The other warning was for *maybe* uninitialized (-Wmaybe-uninitialized)
hb-ot-layout.c. The *maybe* uninitialized variables were changed to be
initialized to zero, though I was unable to find a path where they were
uninitialized. I chose this over silencing the warning for this file.

Test: Compile only.
Change-Id: I2f3faaea536687976391d66baef3a1f7bdcb8141